v0.2.1 Mentors link with activities are fixed, respective activities are returned as they should when the user is a mentor;
