+++ src/main/java/com/solid/solidbackend/entities/Activity.java
@@ -5,6 +5,9 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.util.HashSet;
+import java.util.Set;
+
 @Entity
 @Getter
 @Setter
@@ -21,6 +24,13 @@ public class Activity {
     private User creator;
 
     private String deadline;
+//
+//    @ManyToMany(mappedBy = "activities")
+//    private Set<Team> teams = new HashSet<>();
+//
+//    @ManyToMany(mappedBy = "activities")
+//    private Set<User> mentors = new HashSet<>();
+
 
     public Activity() {
     }
@@ -31,10 +41,27 @@ public class Activity {
         this.deadline = deadline;
     }
 
+//    public Activity(String name, User creator, String deadline, Set<Team> teams, Set<User> mentors) {
+//        this.name = name;
+//        this.creator = creator;
+//        this.deadline = deadline;
+//        this.teams = teams;
+//        this.mentors = mentors;
+//    }
+
     public Activity(Long id, String name, User creator, String deadline) {
         this.id = id;
         this.name = name;
         this.creator = creator;
         this.deadline = deadline;
     }
+
+//    public Activity(Long id, String name, User creator, String deadline, Set<Team> teams, Set<User> mentors) {
+//        this.id = id;
+//        this.name = name;
+//        this.creator = creator;
+//        this.deadline = deadline;
+//        this.teams = teams;
+//        this.mentors = mentors;
+//    }
 }
